sha256 hashing
A function that computes the SHA256 hash of a file, and returns the
first 8-bytes as a uint64_t in identical ordering on both big and little
endian macahines.

Results should be consistent across any machine, so long as no math is
done with the values. As this is intended for uuids, it's a none issue